custom type: structures
